Clippy fixes

+++ lego-powered-up/src/iodevice/basic.rs
@@ -22,7 +22,7 @@ pub trait Basic: Debug + Send + Sync {
         let msg =
             NotificationMessage::PortInputFormatSetupSingle(InputSetupSingle {
                 port_id: self.port(),
-                mode: mode as u8,
+                mode: mode,
                 delta,
                 notification_enabled,
             });
