+++ lego-powered-up/src/iodevice/sensor.rs
@@ -40,7 +40,7 @@ pub trait GenericSensor: Debug + Send + Sync {
         let msg =
             NotificationMessage::PortInputFormatSetupSingle(InputSetupSingle {
                 port_id: self.port(),
-                mode: mode as u8,
+                mode: mode,
                 delta,
                 notification_enabled,
             });
